requirements.txt
Flask==3.0.0
python-telegram-bot==13.15
requests